Advanced and hardcore line optimization strategy
