+++ app/src/main/java/ch/epfl/culturequest/backend/tournament/apis/TournamentManagerApi.java
@@ -1,13 +1,13 @@
 package ch.epfl.culturequest.backend.tournament.apis;
 
 import static androidx.test.core.app.ApplicationProvider.getApplicationContext;
-import static ch.epfl.culturequest.database.Database.fetchTournamentWhenGenerated;
 import static ch.epfl.culturequest.database.Database.handleFutureTimeout;
 import static ch.epfl.culturequest.database.Database.indicateTournamentNotGenerated;
 import static ch.epfl.culturequest.database.Database.isTournamentGenerationLocked;
 import static ch.epfl.culturequest.database.Database.lockTournamentGeneration;
 import static ch.epfl.culturequest.database.Database.unlockTournamentGeneration;
 import static ch.epfl.culturequest.database.Database.uploadTournamentToDatabase;
+import static ch.epfl.culturequest.database.Database.waitForTournamentGenerationAndFetchIt;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
@@ -33,7 +33,11 @@ import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
+import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
@@ -116,6 +120,25 @@ public class TournamentManagerApi {
         return future;
     }
 
+    // If the tournament generation is locked, another device is currently generating the tournament so we should wait for it to be generated and fetch it from the database
+    public static CompletableFuture<Tournament> fetchTournamentWhenGenerated(String tournamentId) {
+
+        AtomicReference<Tournament> fetchedTournament = new AtomicReference<>();
+        CompletableFuture<Tournament> future = new CompletableFuture<>();
+
+        ScheduledExecutorService executor = Executors.newScheduledThreadPool(1);
+        executor.schedule(() -> {
+            if (fetchedTournament.get() == null) {
+                // If the tournament hasn't been generated 2 minutes after the generation lock, it's likely that the device that was generating it crashed so we should unlock the generation and try to generate it again
+                // we make the future complete exceptionally so that the caller (generateOrFetchTournament) can handle it and try to generate the tournament again if needed
+                future.completeExceptionally(new TimeoutException("Failed to fetch the tournament from the database after 2 minutes"));
+            }
+        }, 2, TimeUnit.MINUTES);
+
+        return waitForTournamentGenerationAndFetchIt(fetchedTournament, future, tournamentId);
+    }
+
+
     private static void generateAndStoreTournamentDate() {
 
 
@@ -173,19 +196,14 @@ public class TournamentManagerApi {
 
     }
 
-
     // If the tournament has not been generated yet, generate it and upload it to the database
     // If the tournament has already been generated, fetch it from the database and return it
-
     private static CompletableFuture<Tournament> generateOrFetchTournament() {
 
         return isTournamentGenerationLocked().thenCompose(generationLocked -> {
-
             if (generationLocked) {
-
                 String weeklyTournamentId = RandomApi.getWeeklyTournamentPseudoRandomUUID();
 
-                // wait for the tournament to be generated by another user then fetch it from database
                 return fetchTournamentWhenGenerated(weeklyTournamentId)
                         .handle((result, ex) -> {
                             if (ex != null && ex.getCause() instanceof TimeoutException) {
@@ -205,21 +223,15 @@ public class TournamentManagerApi {
                             } else {
                                 return CompletableFuture.completedFuture(result);
                             }
-                        })
-                        .thenCompose(Function.identity());
-
-            } else {
-
-                // If no one is currently generating the tournament, take the lead and generate it
+                        }).thenCompose(Function.identity());
+            }
+            else {
                 // lock the tournament generation to prevent other users from generating it at the same time
                 lockTournamentGeneration();
 
                 //// Generate the tournament ////
-
-                // pseudo randomly choose the art names to be included in the tournament
                 ArrayList<String> artNames = randomlyChooseArtNames();
 
-                // given the art names, generate the art quizzes
                 Map<String, CompletableFuture<ArtQuiz>> artQuizFutures = generateTournamentQuizzesGivenArtNames(artNames);
 
                 CompletableFuture<ArtQuiz>[] futuresArray = artQuizFutures.values().toArray(new CompletableFuture[0]);
@@ -293,14 +305,12 @@ public class TournamentManagerApi {
     }
 
     private static ArrayList<String> randomlyChooseArtNames() {
-
         Random random = RandomApi.getRandom();
-        ArrayList<String> artNames = new ArrayList<>();
 
+        ArrayList<String> artNames = new ArrayList<>();
         try {
             Resources resources = getApplicationContext().getResources();
             InputStream inputStream = resources.openRawResource(R.raw.famous_arts);
-
             // InputStream to JSONArray
             String json;
             try (Scanner s = new Scanner(inputStream)) {
@@ -321,8 +331,7 @@ public class TournamentManagerApi {
                     String artName = artworkObject.getString("artName");
                     artNames.add(artName);
                     chosenIndices.add(randomIndex);
-                }
-            }
+                }}
         } catch (JSONException e) {
             e.printStackTrace();}
         return artNames;
